+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - Added partial support for CKEditor 4.
 - Added the “CKEditor Build URL” plugin setting, making it possible to customize which CKEditor build is used.
 - Added the “Initialization Code” field setting, making it possible to take control over the editor initialization and configuration.
+- Added `craft\ckeditor\Field::EVENT_MODIFY_PURIFIER_CONFIG`, which makes it possible to modify the HTML Purifier config at runtime.
 
 ### Changed
 - The plugin now requires Craft 3.6 or later.
 - CKEditor 5 (27.0.0, “classic” build) is now used by default.
+- The default HTML Purifier config now allows video embeds from YouTube and Vimeo.
 
 ## 1.0.0-beta.3 - 2020-03-27